[MOV] purchase_alternative*: separate tests files
*: purchase_requisition

With This Commit:
=================================

We have separated the code that depends on the alternative logic
of a tests files from the purchase_requisition module and moved it
into the purchase_alternative module.

taskId :- 4737059
